Separated Logic from UI
Separated the logic from the UI component for clarity.
