+++ ComposureWatch/src/Components/Sorting/SortingLogic.jsx
@@ -1,83 +1,29 @@
 // To be determine if ranks will be part of the rank icon list or housed seperately here for values
 import CardList from "../Card/CardList";
 import shuffle from "./shuffle";
+import teamSorting from "./teamSorting";
 
 const SortingLogic = (props) => {
-  // Logic instructions:
-  // Randomize ids for new teams with same players on "kumite" button press.
-  // Sort players by rating into two objects
-  // Each Team cannot exceed six players
-  // If there is a potential for closer team ratings, having an equal number of players on each team is not a priority.
-  // Good luck  -Zen
-
-  let teamOne = [];
-  let teamTwo = [];
-  let teamOneVal = 0;
-  let teamTwoVal = 0;
-
-  //Fischer-Yates shuffle aglorithm to randomize order each render.  This one is a famous shuffling method I did not create.  Props to the creators.
-
-  shuffle(props.info);
-
-  // Most Basic Test logic for testing purposes
-  //Modify function to account for team rating difference
-  // For example .5 diff max
-  for (let player of props.info) {
-    if (teamTwoVal + player.rating >= teamOneVal) {
-      teamOneVal += player.rating;
-      teamOne.push(player);
-    } else {
-      teamTwoVal += player.rating;
-      teamTwo.push(player);
-    }
-  }
-
-  // Ensures that there is not more than a one player advantage
-  if (teamTwo.length - teamOne.length >= 2) {
-    teamTwoVal -= popPlayerTwo.rating;
-    teamOneVal += popPlayerTwo.rating;
-    teamOne.push(popPlayerTwo);
-  } else if (teamOne.length - teamTwo.length >= 2) {
-    let popPlayerOne = teamOne.pop();
-    teamOneVal -= popPlayerOne.rating;
-    teamTwoVal += popPlayerOne.rating;
-    teamTwo.push(popPlayerOne);
-  }
-
-  // Ensures that teams do not exceed six players
-  if (teamOne.length > 6) {
-    let popPlayerOne = teamOne.pop();
-    teamOneVal -= popPlayerOne.rating;
-    teamTwoVal += popPlayerOne.rating;
-    teamTwo.push(popPlayerOne);
-  } else if (teamTwo.length > 6) {
-    let popPlayerTwo = teamOne.pop();
-    teamTwoVal -= popPlayerTwo.rating;
-    teamOneVal += popPlayerTwo.rating;
-    teamOne.push(popPlayerTwo);
-  }
-
-  let teamOneLength = teamOne.length;
-  let teamTwoLength = teamTwo.length;
+  let teamData = teamSorting(props);
 
   return (
     <>
       <div>
-        <CardList cards={teamOne} team={"teamOne"} />
+        <CardList cards={teamData.teamOne} team={"teamOne"} />
       </div>
       <div className="flex flex-col md:flex-row justify-center items-center py-6 md:space-x-52">
         <div className="text-[#F5122A] text-xl">
-          <p>Team One Value: {teamOneVal.toFixed(2)}</p>
-          <p>Players: {teamOneLength}</p>
+          <p>Team One Value: {teamData.teamOneVal.toFixed(2)}</p>
+          <p>Players: {teamData.teamOneLength}</p>
         </div>
         <div className="text-white font-bold italic text-7xl py-2">VS</div>
         <div className="text-[#2EF3FF] text-xl">
-          <p>Team Two Value: {teamTwoVal.toFixed(2)}</p>
-          <p>Players: {teamTwoLength}</p>
+          <p>Team Two Value: {teamData.teamTwoVal.toFixed(2)}</p>
+          <p>Players: {teamData.teamTwoLength}</p>
         </div>
       </div>
       <div>
-        <CardList cards={teamTwo} team={"teamTwo"} />
+        <CardList cards={teamData.teamTwo} team={"teamTwo"} />
       </div>
     </>
   );
